Renaming 'visualizations' to 'search'
