+++ usaspending_api/search/v2/urls_search.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from usaspending_api.search.v2.views.visualizations import SpendingOverTimeVisualizationViewSet
+from usaspending_api.search.v2.views.search import SpendingOverTimeVisualizationViewSet
 
 urlpatterns = [
     url(r'^spending_over_time$', SpendingOverTimeVisualizationViewSet.as_view())
